requirements.txt
openai-agents>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.10.0


